Matrix class to .h file
